change dbfile

+++ src/main/java/com/yseon/scoremanager/ScoreList.java
@@ -3,22 +3,22 @@ package com.yseon.scoremanager;
 public class ScoreList {
 	int idx;
 	String name;
-	int midScore;
-	int finScore;
+	double midScore;
+	double finScore;
 	String created;
 	
 	ScoreList(){
 		
 	}
 	
-	ScoreList(String name, int mid, int fin, String created){
+	ScoreList(String name, double mid, double fin, String created){
 		this.name = name;
 		this.midScore = mid;
 		this.finScore = fin;
 		this.created = created;
 	}
 	
-	ScoreList(int idx, String name, int mid, int fin, String created) {
+	ScoreList(int idx, String name, double mid, double fin, String created) {
 		this.idx = idx;
 		this.name = name;
 		this.midScore = mid;
